Update SQL QUERY-5.sql
Here we have to listing out all the student id’s who have either joined the institute in the year 2016 or were enrolled for CS101.
So, the correct query is "joining year is 2016 OR course is CS101" and for output we need only student id column i.e. "SELECT stud_id".